+++ packages/s2-core/__tests__/spreadsheet/interaction-brush-selection-scroll-spec.ts
@@ -52,7 +52,7 @@ const dataCfg: S2DataConfig = {
 const options: S2Options = {
   width: 800,
   height: 600,
-  hdAdapter: false,
+  hd: false,
   seriesNumber: {
     enable: true,
   },