+++ packages/s2-react/playground/config.tsx
@@ -328,7 +328,7 @@ export const s2Options: SheetComponentOptions = {
     hoverAfterScroll: true,
     selectedCellMove: true,
     rangeSelection: true,
-    // 防止 mac 触摸板横向滚动触发浏览器返回, 和移动端下拉刷新
+    // 防止 mac 触控板横向滚动触发浏览器返回, 和移动端下拉刷新
     overscrollBehavior: 'none',
     brushSelection: {
       dataCell: true,